Refactor error handling into separate function
